refactor: 최상위 폴더를 가져오는 메서드와 하위 노드를 가져오는 메서드의 위치를 FolderController로 변경
