+++ server/src/main/java/wap/starlist/bookmark/controller/BookmarkController.java
@@ -105,16 +105,4 @@ public class BookmarkController {
             return ResponseEntity.badRequest().body(e.getMessage());
         }
     }
-
-    @GetMapping("/children/{id}")
-    public ResponseEntity<?> getChildren(@PathVariable("id") Long id) {
-        List<BookmarkNodeResponse> nodes = bookmarkService.getChildrenOfFolder(id);
-        return ResponseEntity.ok().body(nodes);
-    }
-
-    @GetMapping("/top-folders")
-    public ResponseEntity<?> getChildFolders(@AuthenticationPrincipal String loginUser) {
-        List<BookmarkNodeResponse> childFolders = bookmarkService.getChildrenOfRoot(loginUser);
-        return ResponseEntity.ok().body(childFolders);
-    }
 }
