+++ server/src/main/java/wap/starlist/bookmark/service/FolderService.java
@@ -1,16 +1,27 @@
 package wap.starlist.bookmark.service;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
+import wap.starlist.bookmark.domain.Bookmark;
 import wap.starlist.bookmark.domain.Folder;
+import wap.starlist.bookmark.domain.Root;
+import wap.starlist.bookmark.dto.response.BookmarkNodeResponse;
 import wap.starlist.bookmark.repository.FolderRepository;
+import wap.starlist.bookmark.repository.RootRepository;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class FolderService {
 
     private final FolderRepository folderRepository;
+    private final RootRepository rootRepository;
 
     public Folder createFolder(String title, Long userId) {
         // 필수값 검증(title, userId 존재 여부 확인)
@@ -40,4 +51,40 @@ public class FolderService {
         Folder folder = getFolder(id);
         folderRepository.delete(folder);
     }
+
+    /**
+     *
+     * @param id: jpa에서 자동으로 생성되는 id값을 의미한다. 이를 통해 member에 해당하는 폴더를 찾기 위해 불필요한 쿼리를 작성하지 않아도 된다.
+     * @return
+     */
+    public List<BookmarkNodeResponse> getChildrenOfFolder(Long id) {
+        Optional<Folder> found = folderRepository.findById(id);
+
+        if (found.isPresent()) {
+            List<Bookmark> bookmarks = found.get().getBookmarks();
+            List<Folder> folders = found.get().getFolders();
+            List<BookmarkNodeResponse> nestedNodes = new ArrayList<>();
+
+            //TODO: parentId가 null인 오류 수정
+            for (Bookmark bookmark : bookmarks) {
+                nestedNodes.add(BookmarkNodeResponse.fromBookmark(bookmark));
+            }
+            for (Folder folder : folders) {
+                nestedNodes.add(BookmarkNodeResponse.fromFolder(folder));
+            }
+
+            return nestedNodes;
+        }
+        log.warn("폴더의 하위 노드가 존재하지 않음: id={}", id);
+        return Collections.emptyList();
+    }
+
+    public List<BookmarkNodeResponse> getChildrenOfRoot(String memberProviderId) {
+        Optional<Root> found = rootRepository.findByMemberProviderId(memberProviderId);
+
+        return found.map(root -> root.getFolders().stream()
+                        .map(BookmarkNodeResponse::fromFolder).toList())
+                .orElse(Collections.emptyList());
+    }
+
 }
